<rev> Improving error reporting, cleaning up MPEG ioctl interface
git-svn-id: file:///export/repository/projects/pchdtv/libtuner@216 e29b0fa8-351f-db11-9190-0011d8377dfc

+++ dvb_driver.h
@@ -80,7 +80,7 @@ typedef struct
 typedef struct
 {
    dvb_modulation_t modulation;
-   uint32_t frequency_hz;
+   uint64_t frequency_hz;
    uint32_t bandwidth_hz;
 } dvb_channel;
